Dont show node with diff in only docs when nodes or tree diff is selected
