+++ src/dotnet/APIView/ClientSPA/src/app/_components/code-panel/code-panel.component.ts
@@ -1,12 +1,11 @@
 import { ChangeDetectorRef, Component, EventEmitter, Input, OnChanges, Output, SimpleChanges } from '@angular/core';
 import { take } from 'rxjs/operators';
-import { CodePanelData } from 'src/app/_models/revision';
 import { Datasource, IDatasource, SizeStrategy } from 'ngx-ui-scroll';
 import { CommentsService } from 'src/app/_services/comments/comments.service';
 import { getQueryParams } from 'src/app/_helpers/router-helpers';
 import { ActivatedRoute, Router } from '@angular/router';
 import { SCROLL_TO_NODE_QUERY_PARAM } from 'src/app/_helpers/common-helpers';
-import { CodePanelRowData, CodePanelRowDatatype } from 'src/app/_models/codePanelRowData';
+import { CodePanelData, CodePanelRowData, CodePanelRowDatatype } from 'src/app/_models/codePanelModels';
 import { StructuredToken } from 'src/app/_models/structuredToken';
 import { CommentItemModel, CommentType } from 'src/app/_models/commentItemModel';
 import { UserProfile } from 'src/app/_models/userProfile';
