+++ src/dotnet/APIView/ClientSPA/src/app/_components/shared/comment-thread/comment-thread.component.spec.ts
@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { CommentThreadComponent } from './comment-thread.component';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { SharedAppModule } from 'src/app/_modules/shared/shared-app.module';
-import { CodePanelRowData } from 'src/app/_models/codePanelRowData';
+import { CodePanelRowData } from 'src/app/_models/codePanelModels';
 import { ReviewPageModule } from 'src/app/_modules/review-page/review-page.module';
 
 describe('CommentThreadComponent', () => {
