+++ src/dotnet/APIView/ClientSPA/src/app/_models/insertCodePanelRowDataMessage.ts
@@ -1,0 +1,11 @@
+export enum ReviewPageWorkerMessageDirective {
+  CreatePageNavigation,
+  UpdateCodePanelData,
+  UpdateCodePanelRowData,
+  SetHasHiddenAPIFlag
+}
+
+export interface InsertCodePanelRowDataMessage {
+  directive: ReviewPageWorkerMessageDirective
+  payload : any
+}
