+++ src/dotnet/APIView/ClientSPA/src/app/_models/navigationTreeModels.ts
@@ -1,0 +1,12 @@
+export class NavigationTreeNodeData {
+  nodeIdHashed: string = '';
+  kind: string = '';
+  icon: string = '';
+}
+  
+export class NavigationTreeNode {
+  label: string = '';
+  data: NavigationTreeNodeData = new NavigationTreeNodeData();
+  expanded: boolean = false;
+  children: NavigationTreeNode[] = [];
+}
