+++ src/dotnet/APIView/ClientSPA/src/app/_services/reviews/reviews.service.ts
@@ -4,7 +4,7 @@ import { Injectable } from '@angular/core';
 import { Observable, map } from 'rxjs';
 import { PaginatedResult } from 'src/app/_models/pagination';
 import { Review } from 'src/app/_models/review';
-import { APIRevision, CodePanelData } from 'src/app/_models/revision';
+import { APIRevision } from 'src/app/_models/revision';
 import { ConfigService } from '../config/config.service';
 
 @Injectable({
